Frame collision prevention, fix u8.

+++ src/stack/router.rs
@@ -13,23 +13,23 @@ use crate::stack::message::{BroadcastMessage, IPAssignFailureMessage};
 
 #[derive(Clone)]
 pub struct MeshRouter {
-    nodeid: i32,
+    nodeid: u8,
     nodeipaddr: Option<Ipv4Addr>,
     gatewayipaddr: Option<Ipv4Addr>,
-    maxhops: i32,
-    lastSequenceNumber: i32,
+    maxhops: u8,
+    lastSequenceNumber: u8,
     timeout: Duration,
-    retries: i32,
-    observations: RefCell<HashMap<i32, Instant>>,
-    graph: UnGraphMap<i32, i32>,
-    id2ip: RefCell<HashMap<i32, Ipv4Addr>>,
-    ip2id: RefCell<HashMap<Ipv4Addr, i32>>,
+    retries: u8,
+    observations: RefCell<HashMap<u8, Instant>>,
+    graph: UnGraphMap<u8, u8>,
+    id2ip: RefCell<HashMap<u8, Ipv4Addr>>,
+    ip2id: RefCell<HashMap<Ipv4Addr, u8>>,
     isgateway: bool
 
 }
 
 impl MeshRouter {
-    pub fn new(nodeid: i32, nodeipaddr: Option<Ipv4Addr>, gatewayipaddr: Option<Ipv4Addr>, maxhops: i32, timeout: Duration, isgateway: bool) -> Self {
+    pub fn new(nodeid: u8, nodeipaddr: Option<Ipv4Addr>, gatewayipaddr: Option<Ipv4Addr>, maxhops: u8, timeout: Duration, isgateway: bool) -> Self {
         let mut router = MeshRouter{
             nodeid,
             nodeipaddr,
@@ -62,7 +62,7 @@ impl MeshRouter {
     }
 
     /// Adds a new route to the mesh, fail if route does not exist
-    pub fn route_add(&mut self, route: Vec<(i32, i32)>) {
+    pub fn route_add(&mut self, route: Vec<(u8, u8)>) {
         route.iter().for_each( |(src, dest)| {
             // we track each observation of every node
             self.node_observe_put(src.clone());
@@ -78,7 +78,7 @@ impl MeshRouter {
     }
 
     /// Handle a network broadcast, maybe node needs an IP?
-    pub fn handle_broadcast(&mut self, broadcast: Box<BroadcastMessage>, route: Vec<i32>) -> Result<Option<(Ipv4Addr, bool)>, IPAssignFailureMessage> {
+    pub fn handle_broadcast(&mut self, broadcast: Box<BroadcastMessage>, route: Vec<u8>) -> Result<Option<(Ipv4Addr, bool)>, IPAssignFailureMessage> {
         let srcid = broadcast.header.expect("Broadcast did not have a frame header.").sender();
         if broadcast.isgateway && srcid != self.nodeid {
             info!("New gateway {} discovered with IP {}", &srcid, &broadcast.ipaddr.expect("Gateways must broadcast their IP"));
@@ -106,10 +106,10 @@ impl MeshRouter {
 
     /// Assign IP address to node
     // TODO implement proper DHCP later
-    fn ip_assign(&mut self, nodeid: i32) -> Result<(Ipv4Addr, bool), IPAssignFailureMessage> {
+    fn ip_assign(&mut self, nodeid: u8) -> Result<(Ipv4Addr, bool), IPAssignFailureMessage> {
         match self.id2ip.get_mut().get(&nodeid) {
             None => {
-                let ipaddr = Ipv4Addr::new(172,16,0, nodeid as u8);
+                let ipaddr = Ipv4Addr::new(172,16,0, nodeid);
                 self.id2ip.get_mut().insert(nodeid, ipaddr);
                 self.ip2id.get_mut().insert(ipaddr, nodeid);
                 return Ok((ipaddr, true));
@@ -121,30 +121,30 @@ impl MeshRouter {
     }
 
     /// Track each node observation for routing purposes
-    fn node_observe_put(&mut self, nodeid: i32) {
+    fn node_observe_put(&mut self, nodeid: u8) {
         self.observations.borrow_mut().insert(nodeid, Instant::now());
     }
 
-    pub fn node_observe_get(&mut self, nodeid: &i32) -> Option<&Instant> {
+    pub fn node_observe_get(&mut self, nodeid: &u8) -> Option<&Instant> {
         self.observations.get_mut().get(nodeid)
     }
 
-    fn edge_add(&mut self, src: i32, dest: i32) {
+    fn edge_add(&mut self, src: u8, dest: u8) {
         self.graph.add_edge(src.clone(), dest.clone(), 1);
     }
 
     /// Add a new node to our mesh
-    fn node_add(&mut self, nodeid: i32) {
+    fn node_add(&mut self, nodeid: u8) {
         self.graph.add_node(nodeid);
     }
 
     /// Removes a node from the mesh
-    pub fn node_remove(&mut self, nodeid: i32) {
+    pub fn node_remove(&mut self, nodeid: u8) {
         self.graph.borrow_mut().remove_node(nodeid);
     }
 
     /// Routes an IP packet to a node in the mesh, if it's possible
-    pub fn packet_route(&mut self, packet: &Packet<Vec<u8>>) -> Option<Vec<i32>> {
+    pub fn packet_route(&mut self, packet: &Packet<Vec<u8>>) -> Option<Vec<u8>> {
         trace!("Routing packet from {} to {}", &packet.source(), &packet.destination());
 
         // look up ip and ensure it's in our mesh
